refactor(termui_pty)!: decouple PTY transport from terminal rendering
This commit overhauls the terminal view architecture to strictly adhere to MVVM principles. `TerminalView` is now a pure, testable UI widget that accepts a `VirtualTerminal` state model. `PseudoTerminalView` replaces the old `PlatformView` as the convenience wrapper for wiring `pty2` subprocesses.

Additional changes in this commit:
* perf(termui): bypass string allocations in `Buffer.writeString` via 1D array blitting and `isWideCodePoint` for a 4.4x layout throughput increase.
* perf(termui_pty): remove synchronous, blocking file I/O from `AnsiParser`.
* fix(termui): resolve memory leak in `SceneManager` by cancelling `_debugTouchTimer` on disposal.
* fix(termui_pty): strictly gate mouse reporting behind DECSET `?1000h` and `?1006h`, and fix related integration tests.
* docs(termui_pty): add missing pub.dev hygiene files (README.md, CHANGELOG.md, LICENSE).
* chore(termui_pty): bump package version to 0.2.0-wip.

BREAKING CHANGE: `PlatformView` has been removed. Use `TerminalView` for raw, decoupled terminal rendering, or `PseudoTerminalView` if you need the package to automatically manage and wire a `pty2.PseudoTerminal` instance.
