+++ packages/termui_pty/example/termui_pty.dart
@@ -20,7 +20,7 @@ void main() async {
       final runner = PromptRunner(
         terminal: terminal,
         alternateScreen: true,
-        widget: SizedBox.expand(child: PlatformView(pty: pty)),
+        widget: SizedBox.expand(child: PseudoTerminalView(pty: pty)),
       );
 
       // Run until the user types 'q' in top, which exits top and closes the pty?
