+++ packages/termui_pty/lib/src/ansi_parser.dart
@@ -1,4 +1,3 @@
-
 /// Interface for receiving parsed ANSI terminal sequences.
 abstract class TerminalHandler {
   /// Print raw printable text.
