+++ packages/termui_pty/lib/termui_pty.dart
@@ -1,6 +1,7 @@
 library;
 
-export 'src/platform_view.dart';
+export 'src/pseudo_terminal_view.dart';
+export 'src/terminal_view.dart';
 export 'src/virtual_terminal.dart';
 export 'src/ansi_parser.dart';
 export 'src/input_encoder.dart';
